+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is used for recording Invader's changes. This changelog is based on
 
 ## [Untagged]
 ### Fixed
+- invader-extract: Fixed extension being missing from extracted tags
 - invader-script: Fixed object_name not being convertible to object_list or
   object
 - invader-script: Fixed globals for object definitions not being recognized